Add the hexdump module
It is a debug module. It is made for this project and for other
projects. It prints a hex dump of a given buffer to the standard
output in a working program without a debugger in a hexdump (linux
program) manner.

Now, we can see what the program is going to write or look at binary
headers to see whether they are filled correctly.