Convert all saved number fields to the big-endian order
On different systems compilation of program may cause another
endianness when saving the data from structures as fields.
Such a program will not read or write correct data.

Now, the order of saved bytes is determined exactly for program
on any system - it is big-endian.

Although another endianness like middle-endian, mixed-endian or
bi-endian aren't considered because they are rare.