Add a test for set value to 100Tb in bignumber unit-tests
This big number module is used for file offsets in this project, so
this specific test doesn't concern with the big number. This test
matters only in this project. In general, big numbers should use long
arithmetics and should accept very big numbers with both signs (not
just positive terabytes).